MultipleBet functionality added. Needs to be linked to the MainLoggedGUI

+++ src/main/java/domain/Movement.java
@@ -11,13 +11,13 @@ public class Movement {
 	@Id
 	private int movID;
 	private String movementType;
-	private int money;
+	private float money;
 	private int balance;
 	private User user;
 	private Event event;
 	private Question question;
 	
-	public Movement(int pMovID, int pMoney, String pMovType, User pUser, Event pEvent, Question pQuestion) {
+	public Movement(int pMovID, float pMoney, String pMovType, User pUser, Event pEvent, Question pQuestion) {
 		super();
 		this.movID = pMovID;
 		this.movementType = pMovType;
@@ -67,7 +67,7 @@ public class Movement {
 		this.question = question;
 	}
 
-	public int getMoney() {
+	public float getMoney() {
 		return money;
 	}
